+++ frontend/src/common/configureEpics.js
@@ -3,7 +3,6 @@ import { combineEpics } from 'redux-observable';
 import appEpics from './app/epics';
 import authEpics from './auth/epics';
 import logEpics from './log/epics';
-import bindCredentials from '../common/apis/bindCredentials';
 
 const epics = [
   ...appEpics,
@@ -15,7 +14,6 @@ const configureEpics = deps => (action$, { getState }) =>
   combineEpics(...epics)(action$, {
     ...deps,
     getState,
-    bindCredentials: bindCredentials(getState),
   });
 
 export default configureEpics;