+++ frontend/src/common/errors/index.js
@@ -1,5 +1,8 @@
-export { default as HttpError } from './HttpError';
-export { default as NetworkError } from './NetworkError';
-export { default as PermissionError } from './PermissionError';
+export { default as apiEpicErrorHandle } from './apiEpicErrorHandle';
+export { default as BadRequestError } from './BadRequestError';
 export { default as ContentTypeError } from './ContentTypeError';
+export { default as HttpError } from './HttpError';
 export { default as InvalidDataError } from './InvalidDataError';
+export { default as NetworkError } from './NetworkError';
+export { default as NotFoundError } from './NotFoundError';
+export { default as PermissionError } from './PermissionError';